fix: refactoring the introForElements function + adding tests
